ldap,mysql,pgsql: use bash pipefail option so that failed dumps are reported as such
This should fix Redmine bug #1340.

This option makes pipelines return as status the value of the last (rightmost)
command to exit with a non-zero status, or zero if all commands exit
successfully. See bash(1) for details.

E.g. this prevents the following from exiting with status 0 (!) if pg_dumpall
fails:

        pg_dumpall | gzip