bup.in: Don't save each directory independantly
index isn't really impacted by this, but running index only once looks
cleaner in the code.

save, on the other hand, shouldn't be called one time for each
directory. this has the effect of creating successive commits that only
contain data from one directory each.
to unify all requested files, we need to specify all the directories on
the one and only call to bup-save.