Move bupdir to [dest] section
Well, let's be logical, here. The bupdir is the local backup
destination, and it's where you store the local index files when backing
up remotely. So, it has nothing to do with the source of the backup!

Plus: in do_bup_ssh_con() in bup.helper.in, set some more comprehensive
error messages for unset variables.
Also, don't verify emptyness for the remote bupdir variable. This can be
empty for local backups. Better check emtpyness for the local bupdir,
which _needs_ to be set.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>