Config for remote bup backups doesn't work
When last changing the config option from remote_dir to remote_bupdir, I
forgot to change it in one place: the output to config file in
bup.helper. This was creating config files that would always try to push
in the remote user's home directory (e.g. since the remote dir was
always empty, it would look like this: user@host: )

Signed-off-by: Gabriel Filion <lelutin@gmail.com>