Implement pull-type backups with "on" type.
"bup on" means to run bup on a remote server and then send back data to
store it locally.

It is a pull-style method of backing up a server.

It would seem like the $type variable suffers from a side effect at some
point and gets set to '0', so let's use another variable name.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>