unquote debug output
this has the weird effect of making every line space- or newline-
separated part of the output as a separate argument to the debug
function.

however, when we quote the output to conserve the newlines intact, we
end up losing all but the first line of output.

so this way it's better than before.. we'd have to patch the debug
function to make it handle new lines better to really fix this the right
way.