Switch default excludes to regexp-based exclusions
The default excludes are "complex" exclusions: they don't directly point
to one and only one file or directory to exclude; so we need to use
regexp-based excludes for the default excludes.

also: removed the duplicity cache from the list. It was probably only
there because I copied the basic exclude list from another helper.