Run duplicity in a C locales environment (Closes: #736280).
This workarounds a bug in duplicity. Quoting the reporter (thanks, Jonathan
Dupart <jonathan@dupart.org>, for the detailed report and initial patch):

The duplicity backend is broken when the gpg ecryption is set up and the
system default locale is not english (or C).

The bug reports on duplicity are here :
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=565398
https://bugs.launchpad.net/duplicity/+bug/510625